Update commentPopup and index.js
